Refactor BookPreview component
Splits the BookPreview component into smaller, more manageable files to improve code organization and maintainability. No functional changes are included.
