+++ src/components/book/BookPreview.tsx
@@ -1,22 +1,14 @@
 
-import { useState, useEffect, useRef } from "react";
+import { useEffect, useRef } from "react";
 import { useQuery } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
-import { Book, ChevronLeft, ChevronRight, X, ZoomIn, ZoomOut, Bookmark } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Skeleton } from "@/components/ui/skeleton";
 import { useToast } from "@/hooks/use-toast";
-import BookCover from "./BookCover";
-import PageView from "./PageView";
-import TableOfContents from "./TableOfContents";
-import { useCoverData } from "@/hooks/useCoverData";
 import { Story } from "@/types/supabase";
-import { cn } from "@/lib/utils";
-import { 
-  calculateTotalPages, 
-  calculateStoryPages, 
-  mapPageToStory 
-} from "@/utils/bookPagination";
+import { useCoverData } from "@/hooks/useCoverData";
+import { useBookNavigation } from "@/hooks/useBookNavigation";
+import BookPreviewHeader from "./BookPreviewHeader";
+import BookPreviewContent from "./BookPreviewContent";
+import TableOfContents from "./TableOfContents";
 
 interface BookPreviewProps {
   profileId: string;
@@ -25,12 +17,6 @@ interface BookPreviewProps {
 }
 
 const BookPreview = ({ profileId, open, onClose }: BookPreviewProps) => {
-  const [currentPage, setCurrentPage] = useState(0);
-  const [zoomLevel, setZoomLevel] = useState(1);
-  const [showToc, setShowToc] = useState(false);
-  const [bookmarks, setBookmarks] = useState<number[]>([]);
-  const [storyPages, setStoryPages] = useState<number[]>([]);
-  const [totalPageCount, setTotalPageCount] = useState(1); // Cover page by default
   const bookContainerRef = useRef<HTMLDivElement>(null);
   const { toast } = useToast();
   const { coverData, isLoading: isCoverLoading } = useCoverData(profileId);
@@ -80,85 +66,23 @@ const BookPreview = ({ profileId, open, onClose }: BookPreviewProps) => {
     enabled: open,
   });
 
-  // Calculate page distribution for stories using our new pagination logic
-  useEffect(() => {
-    if (!stories || stories.length === 0) {
-      setStoryPages([]);
-      setTotalPageCount(1); // Just the cover
-      return;
-    }
-
-    let pageCount = 1; // Start with cover page
-    const pageStartIndices: number[] = [];
-
-    // Calculate starting page for each story
-    stories.forEach((story) => {
-      pageStartIndices.push(pageCount);
-      pageCount += calculateStoryPages(story);
-    });
-
-    setStoryPages(pageStartIndices);
-    setTotalPageCount(calculateTotalPages(stories));
-  }, [stories]);
-
-  // Handle page navigation
-  const goToNextPage = () => {
-    if (currentPage < totalPageCount - 1) {
-      setCurrentPage(currentPage + 1);
-    }
-  };
-
-  const goToPrevPage = () => {
-    if (currentPage > 0) {
-      setCurrentPage(currentPage - 1);
-    }
-  };
-
-  // Find which story is displayed on the current page
-  const getCurrentStory = () => {
-    if (currentPage === 0 || !stories || stories.length === 0) {
-      return null;
-    }
-    
-    const mapping = mapPageToStory(currentPage, stories);
-    return mapping.storyIndex >= 0 ? {
-      story: stories[mapping.storyIndex],
-      pageWithinStory: mapping.pageWithinStory,
-      totalPagesInStory: calculateStoryPages(stories[mapping.storyIndex])
-    } : null;
-  };
-
-  // Handle zoom controls
-  const zoomIn = () => {
-    setZoomLevel(Math.min(2, zoomLevel + 0.1));
-  };
-
-  const zoomOut = () => {
-    setZoomLevel(Math.max(0.5, zoomLevel - 0.1));
-  };
-
-  // Toggle bookmark for current page
-  const toggleBookmark = () => {
-    if (bookmarks.includes(currentPage)) {
-      setBookmarks(bookmarks.filter(b => b !== currentPage));
-    } else {
-      setBookmarks([...bookmarks, currentPage]);
-    }
-  };
-
-  // Jump to specific page from TOC
-  const jumpToPage = (pageIndex: number) => {
-    setCurrentPage(pageIndex);
-    setShowToc(false);
-  };
-
-  // Reset state when dialog is opened
-  useEffect(() => {
-    if (open) {
-      setCurrentPage(0);
-      setZoomLevel(1);
-    }
-  }, [open]);
+  // Use our custom hook for book navigation
+  const {
+    currentPage,
+    zoomLevel,
+    showToc,
+    setShowToc,
+    bookmarks,
+    storyPages,
+    totalPageCount,
+    goToNextPage,
+    goToPrevPage,
+    getCurrentStory,
+    zoomIn,
+    zoomOut,
+    toggleBookmark,
+    jumpToPage
+  } = useBookNavigation(stories, open);
 
   // Handle keyboard navigation
   useEffect(() => {
@@ -182,56 +106,29 @@ const BookPreview = ({ profileId, open, onClose }: BookPreviewProps) => {
 
     window.addEventListener("keydown", handleKeyDown);
     return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [open, currentPage, totalPageCount, onClose]);
+  }, [open, currentPage, totalPageCount, onClose, goToNextPage, goToPrevPage]);
 
   if (!open) return null;
 
   // Get the current story to display
   const currentStoryInfo = getCurrentStory();
+  const authorName = profile ? `${profile.first_name} ${profile.last_name}` : "";
 
   return (
     <div className="fixed inset-0 bg-black/80 z-50 flex flex-col items-center justify-start overflow-hidden">
       {/* Header */}
-      <div className="w-full bg-background p-4 flex items-center justify-between">
-        <div className="flex items-center space-x-4">
-          <Button 
-            variant="outline" 
-            size="sm" 
-            onClick={() => setShowToc(!showToc)}
-          >
-            <Book className="h-4 w-4 mr-2" />
-            Table of Contents
-          </Button>
-          <div className="flex items-center space-x-2">
-            <Button variant="outline" size="icon" onClick={zoomOut}>
-              <ZoomOut className="h-4 w-4" />
-            </Button>
-            <span className="text-sm">{Math.round(zoomLevel * 100)}%</span>
-            <Button variant="outline" size="icon" onClick={zoomIn}>
-              <ZoomIn className="h-4 w-4" />
-            </Button>
-          </div>
-          <Button 
-            variant="outline" 
-            size="sm" 
-            onClick={toggleBookmark}
-            className={cn(
-              bookmarks.includes(currentPage) && "text-amber-500"
-            )}
-          >
-            <Bookmark className="h-4 w-4 mr-2" />
-            {bookmarks.includes(currentPage) ? "Bookmarked" : "Bookmark"}
-          </Button>
-        </div>
-        <div className="flex items-center space-x-2">
-          <span className="text-sm">
-            Page {currentPage + 1} of {totalPageCount}
-          </span>
-          <Button variant="ghost" size="icon" onClick={onClose}>
-            <X className="h-5 w-5" />
-          </Button>
-        </div>
-      </div>
+      <BookPreviewHeader
+        totalPageCount={totalPageCount}
+        currentPage={currentPage}
+        zoomLevel={zoomLevel}
+        showToc={showToc}
+        bookmarks={bookmarks}
+        onToggleToc={() => setShowToc(!showToc)}
+        onZoomIn={zoomIn}
+        onZoomOut={zoomOut}
+        onToggleBookmark={toggleBookmark}
+        onClose={onClose}
+      />
 
       <div className="flex-1 w-full flex overflow-hidden">
         {/* TOC Sidebar */}
@@ -252,62 +149,19 @@ const BookPreview = ({ profileId, open, onClose }: BookPreviewProps) => {
           className="flex-1 h-full flex flex-col items-center justify-center p-4 overflow-auto"
           ref={bookContainerRef}
         >
-          <div 
-            className="relative bg-white shadow-xl rounded-md transition-transform"
-            style={{ 
-              transform: `scale(${zoomLevel})`,
-              transformOrigin: 'center',
-              width: '600px',  // Adjusted for 5x8 aspect ratio (5:8 = 600:960)
-              height: '960px', // 5x8 inch ratio
-              maxHeight: '90vh'
-            }}
-          >
-            {/* Book Pages */}
-            {isStoriesLoading || isCoverLoading ? (
-              <Skeleton className="w-full h-full" />
-            ) : (
-              <>
-                {currentPage === 0 ? (
-                  // Cover Page
-                  <BookCover 
-                    coverData={coverData} 
-                    authorName={profile ? `${profile.first_name} ${profile.last_name}` : ""}
-                  />
-                ) : (
-                  // Content Pages
-                  currentStoryInfo && currentStoryInfo.story && (
-                    <PageView 
-                      story={currentStoryInfo.story} 
-                      pageNumber={currentStoryInfo.pageWithinStory}
-                      totalPagesInStory={currentStoryInfo.totalPagesInStory}
-                    />
-                  )
-                )}
-              </>
-            )}
-
-            {/* Page Turn Buttons */}
-            <div className="absolute inset-0 flex justify-between items-center pointer-events-none">
-              <Button 
-                variant="ghost" 
-                size="icon" 
-                onClick={goToPrevPage}
-                disabled={currentPage === 0}
-                className="h-12 w-12 rounded-full bg-background/80 pointer-events-auto ml-2"
-              >
-                <ChevronLeft className="h-6 w-6" />
-              </Button>
-              <Button 
-                variant="ghost" 
-                size="icon" 
-                onClick={goToNextPage}
-                disabled={currentPage === totalPageCount - 1}
-                className="h-12 w-12 rounded-full bg-background/80 pointer-events-auto mr-2"
-              >
-                <ChevronRight className="h-6 w-6" />
-              </Button>
-            </div>
-          </div>
+          <BookPreviewContent
+            currentPage={currentPage}
+            totalPageCount={totalPageCount}
+            zoomLevel={zoomLevel}
+            stories={stories}
+            isStoriesLoading={isStoriesLoading}
+            isCoverLoading={isCoverLoading}
+            coverData={coverData}
+            authorName={authorName}
+            goToNextPage={goToNextPage}
+            goToPrevPage={goToPrevPage}
+            currentStoryInfo={currentStoryInfo}
+          />
         </div>
       </div>
     </div>
