+++ src/components/book/BookPreviewContent.tsx
@@ -1,0 +1,88 @@
+
+import React from "react";
+import { Skeleton } from "@/components/ui/skeleton";
+import BookCover from "./BookCover";
+import PageView from "./PageView";
+import BookPreviewControls from "./BookPreviewControls";
+import { Story } from "@/types/supabase";
+import { CoverData } from "@/components/cover/CoverTypes";
+import { calculateStoryPages } from "@/utils/bookPagination";
+
+interface BookPreviewContentProps {
+  currentPage: number;
+  totalPageCount: number;
+  zoomLevel: number;
+  stories: Story[] | undefined;
+  isStoriesLoading: boolean;
+  isCoverLoading: boolean;
+  coverData: CoverData;
+  authorName: string;
+  goToNextPage: () => void;
+  goToPrevPage: () => void;
+  currentStoryInfo: {
+    story: Story;
+    pageWithinStory: number;
+    totalPagesInStory: number;
+  } | null;
+}
+
+const BookPreviewContent = ({
+  currentPage,
+  totalPageCount,
+  zoomLevel,
+  stories,
+  isStoriesLoading,
+  isCoverLoading,
+  coverData,
+  authorName,
+  goToNextPage,
+  goToPrevPage,
+  currentStoryInfo,
+}: BookPreviewContentProps) => {
+  return (
+    <div 
+      className="relative bg-white shadow-xl rounded-md transition-transform"
+      style={{ 
+        transform: `scale(${zoomLevel})`,
+        transformOrigin: 'center',
+        width: '600px',  // Adjusted for 5x8 aspect ratio (5:8 = 600:960)
+        height: '960px', // 5x8 inch ratio
+        maxHeight: '90vh'
+      }}
+    >
+      {/* Book Pages */}
+      {isStoriesLoading || isCoverLoading ? (
+        <Skeleton className="w-full h-full" />
+      ) : (
+        <>
+          {currentPage === 0 ? (
+            // Cover Page
+            <BookCover 
+              coverData={coverData} 
+              authorName={authorName}
+            />
+          ) : (
+            // Content Pages
+            currentStoryInfo && currentStoryInfo.story && (
+              <PageView 
+                story={currentStoryInfo.story} 
+                pageNumber={currentStoryInfo.pageWithinStory}
+                totalPagesInStory={currentStoryInfo.totalPagesInStory}
+              />
+            )
+          )}
+        </>
+      )}
+
+      {/* Page Turn Buttons */}
+      <BookPreviewControls
+        currentPage={currentPage}
+        totalPageCount={totalPageCount}
+        goToNextPage={goToNextPage}
+        goToPrevPage={goToPrevPage}
+      />
+    </div>
+  );
+};
+
+export default BookPreviewContent;
